+++ src/mruby_integration/models/colour.cpp
@@ -113,6 +113,7 @@ mrb_Colour_tint(mrb_state* mrb, mrb_value self) -> mrb_value
 
   return mrb_Color_value(mrb, return_colour);
 }
+
 auto
 mrb_Colour_brightness(mrb_state* mrb, mrb_value self) -> mrb_value
 {
@@ -122,11 +123,29 @@ mrb_Colour_brightness(mrb_state* mrb, mrb_value self) -> mrb_value
   mrb_get_args(mrb, "f", &brightness);
 
   if (brightness < -1 || brightness > 1) {
-    mrb_raise(mrb, E_ARGUMENT_ERROR, "Brightness must be within (0.0..1.0)");
+    mrb_raise(mrb, E_ARGUMENT_ERROR, "Brightness must be within (-1.0..1.0)");
   }
 
   auto* return_colour = static_cast<Color*>(malloc(sizeof(Color)));
   *return_colour = ColorBrightness(*colour, brightness);
+
+  return mrb_Color_value(mrb, return_colour);
+}
+
+auto
+mrb_Colour_contrast(mrb_state* mrb, mrb_value self) -> mrb_value
+{
+  mrb_get_self(mrb, self, Color, colour);
+
+  mrb_float contrast;
+  mrb_get_args(mrb, "f", &contrast);
+
+  if (contrast < -1 || contrast > 1) {
+    mrb_raise(mrb, E_ARGUMENT_ERROR, "Contrast must be within (-1.0..1.0)");
+  }
+
+  auto* return_colour = static_cast<Color*>(malloc(sizeof(Color)));
+  *return_colour = ColorContrast(*colour, contrast);
 
   return mrb_Color_value(mrb, return_colour);
 }
@@ -156,6 +175,8 @@ append_models_Colour(mrb_state* mrb)
     mrb, Colour_class, "tint", mrb_Colour_tint, MRB_ARGS_REQ(1));
   mrb_define_method(
     mrb, Colour_class, "brightness", mrb_Colour_brightness, MRB_ARGS_REQ(1));
+  mrb_define_method(
+    mrb, Colour_class, "contrast", mrb_Colour_contrast, MRB_ARGS_REQ(1));
 
   load_ruby_models_colour(mrb);
 }